backend/requirements.txt
# SPDX-FileCopyrightText: 2025 Stanford University
#
# SPDX-License-Identifier: MIT

async_lru==2.0.4
fastapi==0.109.2
firebase_admin==6.4.0
icalendar==5.0.11
numpy==1.26.4
openai==1.11.1
protobuf==4.25.2
pydantic==2.6.1
python-dotenv==1.0.1
python_dateutil==2.8.2
pytz==2023.3.post1
recurring_ical_events==2.1.1
Requests==2.31.0
starlette>=0.36.3
apscheduler==3.10.4
pandas==2.1.4
pytest==8.0.2
pytest-asyncio==0.23.6
uvicorn[standard]
